requirements.txt
streamlit==1.36.0
google-generativeai==0.7.1
python-dotenv==1.0.1
requests==2.32.3 # For fetching content from URLs
beautifulsoup4==4.12.3 # For parsing HTML from fetched URLs